# Added Span support to Webencoders