--- a/shared/Microsoft.Extensions.WebEncoders.Sources/Properties/EncoderResources.cs
+++ b/shared/Microsoft.Extensions.WebEncoders.Sources/Properties/EncoderResources.cs
@@ -20,6 +20,11 @@ namespace Microsoft.Extensions.WebEncoders.Sources
         internal static readonly string WebEncoders_MalformedInput = "Malformed input: {0} is an invalid input length.";
 
         /// <summary>
+        /// Invalid input, that doesn't conform a base64 string.
+        /// </summary>
+        internal static readonly string WebEncoders_InvalidInput = "The input is not a valid Base-64 string as it contains a non-base 64 character, more than two padding characters, or an illegal character among the padding characters.";
+
+        /// <summary>
         /// Invalid {0}, {1} or {2} length.
         /// </summary>
         internal static string FormatWebEncoders_InvalidCountOffsetOrLength(object p0, object p1, object p2)