… benchmarks/Microsoft.Extensions.Internal.Benchmarks/Microsoft.Extensions.Internal.Benchmarks.csproj
… benchmarks/Microsoft.Extensions.Internal.Benchmarks/Properties/AssemblyInfo.cs
… benchmarks/Microsoft.Extensions.Internal.Benchmarks/WebEncodersBenchmarks.cs
… test/Microsoft.Extensions.Internal.Test/Microsoft.Extensions.Internal.Test.csproj
… test/Microsoft.Extensions.Internal.Test/WebEncodersTests.cs