# Description

* Fixes https://github.com/aspnet/Home/issues/2966
* added benchmark-project for `WebEncoders`
* added Span-support to the API (see section _API_) and used the Span-based methods internally
* small buffers (<= 256 bytes length) are stack allocated, larger ones rent from ArrayPool (except net461, there the large ones are heap allocated)
* vectorized the url-encoding / -decoding step
* more or less a rewrite of the class (and so it got a bit longer)
* tried to reach best perf and minimal heap allocations for netcoreapp2.1, netcoreapp2.0, and net461

Due the latter point some `#if`s are needed, so the code doesn't look as nice as it could on some places.

## API

```diff
public static byte[] Base64UrlDecode(string input)
 public static byte[] Base64UrlDecode(string input, int offset, int count)
+public static byte[] Base64UrlDecode(ReadOnlySpan<char> base64Url)
+public static int Base64UrlDecode(ReadOnlySpan<char> base64Url, Span<byte> data)
+public static OperationStatus Base64UrlDecode(ReadOnlySpan<byte> base64Url, Span<byte> data, out int bytesConsumed, out int bytesWritten, bool isFinalBlock = true)
 public static byte[] Base64UrlDecode(string input, int offset, char[] buffer, int bufferOffset, int count)
 public static string Base64UrlEncode(byte[] input)
 public static string Base64UrlEncode(byte[] input, int offset, int count)
+public static string Base64UrlEncode(ReadOnlySpan<byte> data)
+public static int Base64UrlEncode(ReadOnlySpan<byte> data, Span<char> base64Url)
+public static OperationStatus Base64UrlEncode(ReadOnlySpan<byte> data, Span<byte> base64Url, out int bytesConsumed, out int bytesWritten, bool isFinalBlock = true)
 public static int Base64UrlEncode(byte[] input, int offset, char[] output, int outputOffset, int count)
 public static int GetArraySizeRequiredToDecode(int count)
 public static int GetArraySizeRequiredToEncode(int count)
```

# Benchmarks

## Results from Benchmark-Project

### Baseline

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.300-preview3-008416
  [Host]     : .NET Core 2.1.0-preview2-26314-02 (Framework 4.6.26310.01), 64bit RyuJIT
  Job-GJOWXM : .NET Core 2.1.0-preview2-26314-02 (Framework 4.6.26310.01), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 2.1  
RunStrategy=Throughput  

```
|                         Method |          Mean |       Error |      StdDev |        Median |             Op/s |  Gen 0 | Allocated |
|------------------------------- |--------------:|------------:|------------:|--------------:|-----------------:|-------:|----------:|
|           Base64UrlDecode_Data | 5,168.7843 ns | 114.5844 ns | 336.0562 ns | 5,181.8617 ns |        193,469.1 | 0.0301 |    1888 B |
| Base64UrlDecode_DataWithOffset | 5,290.3002 ns | 178.3533 ns | 525.8787 ns | 5,302.4528 ns |        189,025.2 | 0.0302 |    1888 B |
|           Base64UrlDecode_Guid |   237.9833 ns |   6.1025 ns |  17.5093 ns |   234.4482 ns |      4,201,976.3 | 0.0017 |     112 B |
|           Base64UrlEncode_Data | 3,124.1497 ns | 111.2723 ns | 109.2843 ns | 3,079.0000 ns |        320,087.1 | 0.0455 |    2720 B |
| Base64UrlEncode_DataWithOffset | 3,269.8675 ns |  86.4926 ns | 252.3030 ns | 3,197.4447 ns |        305,822.8 | 0.0455 |    2720 B |
|           Base64UrlEncode_Guid |   172.6008 ns |   5.4054 ns |  15.8531 ns |   168.7662 ns |      5,793,717.2 | 0.0024 |     144 B |
|   GetArraySizeRequiredToDecode |     6.6558 ns |   0.2717 ns |   0.8012 ns |     6.5851 ns |    150,244,480.7 |      - |       0 B |
|   GetArraySizeRequiredToEncode |     0.0377 ns |   0.0312 ns |   0.0554 ns |     0.0004 ns | 26,507,091,393.8 |      - |       0 B |

### This PR

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.300-preview3-008416
  [Host]     : .NET Core 2.1.0-preview2-26314-02 (Framework 4.6.26310.01), 64bit RyuJIT
  Job-MVRCXO : .NET Core 2.1.0-preview2-26314-02 (Framework 4.6.26310.01), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 2.1  
RunStrategy=Throughput  

```
|                         Method |          Mean |      Error |      StdDev |        Median |             Op/s |  Gen 0 | Allocated |
|------------------------------- |--------------:|-----------:|------------:|--------------:|-----------------:|-------:|----------:|
|           Base64UrlDecode_Data | 1,293.3192 ns | 57.9273 ns | 168.0575 ns | 1,259.1617 ns |        773,204.3 | 0.0076 |     528 B |
| Base64UrlDecode_DataWithOffset | 1,335.1744 ns | 61.1991 ns | 178.5208 ns | 1,306.9594 ns |        748,965.8 | 0.0076 |     528 B |
|           Base64UrlDecode_Guid |   155.3893 ns |  5.1800 ns |  15.2734 ns |   153.4095 ns |      6,435,450.5 | 0.0005 |      40 B |
|           Base64UrlEncode_Data | 1,584.6087 ns | 13.6427 ns |  10.6513 ns | 1,580.5804 ns |        631,070.6 | 0.0228 |    1360 B |
| Base64UrlEncode_DataWithOffset | 1,621.0625 ns | 29.7223 ns |  23.2052 ns | 1,615.7145 ns |        616,879.3 | 0.0226 |    1360 B |
|           Base64UrlEncode_Guid |   160.5265 ns |  5.6154 ns |  16.5571 ns |   160.0300 ns |      6,229,501.0 | 0.0012 |      72 B |
|   GetArraySizeRequiredToDecode |     0.0530 ns |  0.0317 ns |   0.0784 ns |     0.0001 ns | 18,869,922,106.3 |      - |       0 B |
|   GetArraySizeRequiredToEncode |     0.0764 ns |  0.0409 ns |   0.0853 ns |     0.0531 ns | 13,088,763,160.8 |      - |       0 B |

### Notes

This benchmarks were run on ubuntu-x64. On win-x64 the results should be even better (see next section), but I don't have numbers, because it takes so long....

In `GetArraySizeRequiredToEncode` there is a little slow-down, but this shouldn't care.

## Individual benchmarks 

[Project for this benchmarks](https://github.com/gfoidl/Benchmarks/tree/d2f1010b8bbbbc745ae4f8222a91b2d8c08fd370/aspnet/Common/WebEncoders)

### Decode Guid

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742189 Hz, Resolution=364.6722 ns, Timer=TSC
.NET Core SDK=2.1.300-preview3-008384
  [Host] : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT
  Clr	 : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  Core	 : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT


```
|  Method |	 Job | Runtime |	 Mean |	   Error |	 StdDev | Scaled | ScaledSD |  Gen 0 | Allocated |
|-------- |----- |-------- |---------:|---------:|---------:|-------:|---------:|-------:|----------:|
| Default |	 Clr |	   Clr | 159.8 ns | 3.171 ns | 3.525 ns |	1.00 |	   0.00 | 0.0355 |	   112 B |
|	  New |	 Clr |	   Clr | 126.4 ns | 2.110 ns | 1.973 ns |	0.79 |	   0.02 | 0.0126 |		40 B |
|		  |		 |		   |		  |			 |			|		 |			|		 |			 |
| Default | Core |	  Core | 172.6 ns | 3.249 ns | 2.880 ns |	1.00 |	   0.00 | 0.0355 |	   112 B |
|	  New | Core |	  Core | 119.0 ns | 2.292 ns | 2.032 ns |	0.69 |	   0.02 | 0.0126 |		40 B |

### Decode Data

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742189 Hz, Resolution=364.6722 ns, Timer=TSC
.NET Core SDK=2.1.300-preview3-008384
  [Host] : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT
  Clr	 : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  Core	 : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT


```
|  Method |	 Job | Runtime | DataSize |		Mean |		Error |		StdDev |   Median | Scaled | ScaledSD |	 Gen 0 | Allocated |
|-------- |----- |-------- |--------- |---------:|-----------:|-----------:|---------:|-------:|---------:|-------:|----------:|
| **Default** |	 **Clr** |	   **Clr** |	   **10** | **119.7 ns** |	**2.4100 ns** |	 **2.6787 ns** | **119.4 ns** |	  **1.00** |	 **0.00** | **0.0303** |	  **96 B** |
|	  New |	 Clr |	   Clr |	   10 | 124.7 ns |	2.5604 ns |	 3.4181 ns | 124.7 ns |	  1.04 |	 0.04 | 0.0126 |	  40 B |
|		  |		 |		   |		  |			 |			  |			   |		  |		   |		  |		   |		   |
| Default | Core |	  Core |	   10 | 131.0 ns |	2.7016 ns |	 5.6392 ns | 129.9 ns |	  1.00 |	 0.00 | 0.0303 |	  96 B |
|	  New | Core |	  Core |	   10 | 106.6 ns |	3.3803 ns |	 9.6988 ns | 103.2 ns |	  0.82 |	 0.08 | 0.0126 |	  40 B |
|		  |		 |		   |		  |			 |			  |			   |		  |		   |		  |		   |		   |
| **Default** |	 **Clr** |	   **Clr** |	   **50** | **398.0 ns** |	**7.8665 ns** |	 **9.3645 ns** | **398.5 ns** |	  **1.00** |	 **0.00** | **0.0758** |	 **240 B** |
|	  New |	 Clr |	   Clr |	   50 | 203.1 ns |	1.3501 ns |	 1.0541 ns | 202.9 ns |	  0.51 |	 0.01 | 0.0253 |	  80 B |
|		  |		 |		   |		  |			 |			  |			   |		  |		   |		  |		   |		   |
| Default | Core |	  Core |	   50 | 430.2 ns |	3.4316 ns |	 3.2099 ns | 431.0 ns |	  1.00 |	 0.00 | 0.0758 |	 240 B |
|	  New | Core |	  Core |	   50 | 132.8 ns |	0.7019 ns |	 0.6222 ns | 133.0 ns |	  0.31 |	 0.00 | 0.0253 |	  80 B |
|		  |		 |		   |		  |			 |			  |			   |		  |		   |		  |		   |		   |
| **Default** |	 **Clr** |	   **Clr** |	  **100** | **744.1 ns** | **10.4349 ns** |	 **8.7136 ns** | **742.6 ns** |	  **1.00** |	 **0.00** | **0.1345** |	 **424 B** |
|	  New |	 Clr |	   Clr |	  100 | 380.2 ns |	7.5157 ns | 12.5570 ns | 380.2 ns |	  0.51 |	 0.02 | 0.0911 |	 288 B |
|		  |		 |		   |		  |			 |			  |			   |		  |		   |		  |		   |		   |
| Default | Core |	  Core |	  100 | 866.1 ns | 15.9919 ns | 22.4184 ns | 864.3 ns |	  1.00 |	 0.00 | 0.1345 |	 424 B |
|	  New | Core |	  Core |	  100 | 277.8 ns |	5.5897 ns |	 9.9358 ns | 278.1 ns |	  0.32 |	 0.01 | 0.0405 |	 128 B |


### Encode Guid

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742189 Hz, Resolution=364.6722 ns, Timer=TSC
.NET Core SDK=2.1.300-preview3-008384
  [Host] : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT
  Clr	 : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  Core	 : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT


```
|  Method |	 Job | Runtime |	  Mean |	Error |	  StdDev | Scaled | ScaledSD |	Gen 0 | Allocated |
|-------- |----- |-------- |----------:|---------:|---------:|-------:|---------:|-------:|----------:|
| Default |	 Clr |	   Clr | 143.72 ns | 2.896 ns | 3.448 ns |	 1.00 |		0.00 | 0.0455 |		144 B |
|	  New |	 Clr |	   Clr | 123.75 ns | 2.683 ns | 5.945 ns |	 0.86 |		0.05 | 0.0455 |		144 B |
|		  |		 |		   |		   |		  |			 |		  |			 |		  |			  |
| Default | Core |	  Core | 107.76 ns | 2.183 ns | 2.144 ns |	 1.00 |		0.00 | 0.0457 |		144 B |
|	  New | Core |	  Core |  82.49 ns | 1.721 ns | 2.048 ns |	 0.77 |		0.02 | 0.0228 |		 72 B |


### Encode Data

``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742189 Hz, Resolution=364.6722 ns, Timer=TSC
.NET Core SDK=2.1.300-preview3-008384
  [Host] : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT
  Clr	 : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0
  Core	 : .NET Core 2.1.0-preview2-26313-01 (CoreCLR 4.6.26310.01, CoreFX 4.6.26313.01), 64bit RyuJIT


```
|  Method |	 Job | Runtime | DataSize |		 Mean |		Error |	   StdDev | Scaled | ScaledSD |	 Gen 0 | Allocated |
|-------- |----- |-------- |--------- |----------:|----------:|----------:|-------:|---------:|-------:|----------:|
| **Default** |	 **Clr** |	   **Clr** |	   **10** |	 **99.66 ns** |	 **2.085 ns** |	 **4.259 ns** |	  **1.00** |	 **0.00** | **0.0355** |	 **112 B** |
|	  New |	 Clr |	   Clr |	   10 |	 96.82 ns |	 1.934 ns |	 1.986 ns |	  0.97 |	 0.04 | 0.0355 |	 112 B |
|		  |		 |		   |		  |			  |			  |			  |		   |		  |		   |		   |
| Default | Core |	  Core |	   10 |	 94.31 ns |	 1.966 ns |	 5.178 ns |	  1.00 |	 0.00 | 0.0355 |	 112 B |
|	  New | Core |	  Core |	   10 |	 85.06 ns |	 1.768 ns |	 2.479 ns |	  0.90 |	 0.05 | 0.0178 |	  56 B |
|		  |		 |		   |		  |			  |			  |			  |		   |		  |		   |		   |
| **Default** |	 **Clr** |	   **Clr** |	   **50** | **292.39 ns** |	 **1.660 ns** |	 **1.472 ns** |	  **1.00** |	 **0.00** | **0.1016** |	 **320 B** |
|	  New |	 Clr |	   Clr |	   50 | 242.92 ns |	 4.917 ns | 10.897 ns |	  0.83 |	 0.04 | 0.1016 |	 320 B |
|		  |		 |		   |		  |			  |			  |			  |		   |		  |		   |		   |
| Default | Core |	  Core |	   50 | 240.90 ns |	 4.883 ns |	 4.796 ns |	  1.00 |	 0.00 | 0.1016 |	 320 B |
|	  New | Core |	  Core |	   50 | 151.45 ns |	 3.082 ns |	 5.479 ns |	  0.63 |	 0.03 | 0.0508 |	 160 B |
|		  |		 |		   |		  |			  |			  |			  |		   |		  |		   |		   |
| **Default** |	 **Clr** |	   **Clr** |	  **100** | **557.73 ns** | **10.927 ns** | **13.008 ns** |	  **1.00** |	 **0.00** | **0.1879** |	 **592 B** |
|	  New |	 Clr |	   Clr |	  100 | 431.71 ns |	 7.723 ns |	 6.846 ns |	  0.77 |	 0.02 | 0.1879 |	 592 B |
|		  |		 |		   |		  |			  |			  |			  |		   |		  |		   |		   |
| Default | Core |	  Core |	  100 | 467.78 ns |	 9.396 ns | 19.614 ns |	  1.00 |	 0.00 | 0.1879 |	 592 B |
|	  New | Core |	  Core |	  100 | 326.64 ns |	 6.474 ns |	 5.739 ns |	  0.70 |	 0.03 | 0.0939 |	 296 B |


# Notes

## Prospects

There could be a ~20% perf-win, when the base64-encoding / -decoding step would convert directly to / from url-encoded form by changing the base64-character map.
But by doing so we would "duplicate" the methods from `Base64`-class. This PR does rely on the provided methods -- maybe a follow-up PR will handle this.

## Private classes

In a normal project I would separate the private classes to their own files, but here the `WebEncoders.cs` is a linked file, so I stuffed everything in this file.