@GrabYourPitchforks 2018-04-05
BTW, I should point out that the amount of unsafe code being introduced here makes me a bit nervous. It would be great if there were evidence that real-world applications - not just microbenchmarks - would benefit from this change.