@davidfowl 2018-04-05
> BTW, I should point out that the amount of unsafe code being introduced here makes me a bit nervous. It would be great if there were evidence that real-world applications - not just microbenchmarks - would benefit from this change.

We need to spanify all the things, what can we do to reduce the unsafe code. It's not just about throughput but allocations.