@gfoidl 2018-04-05
> It's not just about throughput but allocations.

So shall I go with (easy) sequential code? Allocations are still minimized, throughput will increase especially on longer inputs.

It's quite a lot of "unsafeness" in there which I don't really like either. I wanted to squeeze everything out, because then going backward (i.e. removing some unsafe parts) seems easier to me than the other way round.

> evidence that real-world applications

Thats nearly alway the point. I don't have any good context where this class is and will be used -- just assumed it as "low-level lib".
Do we have any more real world test-case for this? 
If the unsafeness isn't it worth, then we shouldn't have it in the code and prefer easy to read and maintain alternatives.