@gfoidl 2018-04-07
So I believe this one will take longer :wink:
https://github.com/dotnet/corefx/issues/28114#issuecomment-379410199 gives a clue that 2.1 will be near, and for https://github.com/aspnet/SignalR/pull/1606#discussion_r175005138 it would be nice to have at least
```diff
+public static string Base64UrlEncode(ReadOnlySpan<byte> data)
```

Shall I create a very trimmed down PR that just handles this method (so it can go for 2.1)?
This PR can be continued later.