@gfoidl 2018-04-16
https://github.com/aspnet/Common/pull/338 is a direct way of encoding, and -- at least for me -- a better solution than this PR. So I would close this PR in favor of https://github.com/aspnet/Common/pull/338.