@gfoidl 2018-04-27
Rebased due to conflict in `Common.sln` (`Microsoft.AspNetCore.Analyzer.Testing.csproj` added to `dev`, this PR adds `Microsoft.Extensions.Internal.Benchmarks.csproj`).