@gfoidl 2018-06-18
Rebased due to

```
Conflicting files
Common.sln
```